+++ CHANGELOG.md
@@ -15,6 +15,9 @@ control the live
 - Add a widget to modify live title and allow / disallow live recording
 - Add a widget to schedule live and modify description
 
+### Changed
+
+- Move appairing device component in the widget dashboard
 
 ## [4.0.0-beta.3] - 2022-04-22
 
@@ -25,9 +28,6 @@ control the live
 - Send emails in current livesession's language
 - Load livesession's language in the JWT token for direct access 
 from email
-- Add a control pane destined to contain widgets for the instructor to 
-control the live
-- Add a widget to modify live title and allow / disallow live recording
 
 
 ### Fixed